some mods

+++ libft/ft_lstdelone.c
@@ -6,7 +6,7 @@
 /*   By: svan-der <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/22 15:05:54 by svan-der       #+#    #+#                */
-/*   Updated: 2020/03/07 23:24:47 by svan-der      ########   odam.nl         */
+/*   Updated: 2020/03/09 18:41:27 by svan-der      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ void	content_del(void *content, size_t size)
 	(void)size;
 }
 
-void	ft_lstdelone(t_list **alst, void (*del)(void *, size_t))
+void	ft_lstdelone(t_list **alst, void (*del)(void *, int))
 {
 	if (alst)
 	{
